move old direct etcd methods for policy closer to source
