+++ pkg/cmd/server/admin/legacyetcd/clusterpolicy/etcd/etcd.go
@@ -7,7 +7,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	"github.com/openshift/origin/pkg/authorization/registry/clusterpolicybinding"
+	"github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterpolicy"
 	"github.com/openshift/origin/pkg/util/restoptions"
 )
 
@@ -15,17 +15,17 @@ type REST struct {
 	*registry.Store
 }
 
-// NewREST returns a RESTStorage object that will work against ClusterPolicyBinding.
+// NewREST returns a RESTStorage object that will work against ClusterPolicy.
 func NewREST(optsGetter restoptions.Getter) (*REST, error) {
 	store := &registry.Store{
 		Copier:                   kapi.Scheme,
-		NewFunc:                  func() runtime.Object { return &authorizationapi.ClusterPolicyBinding{} },
-		NewListFunc:              func() runtime.Object { return &authorizationapi.ClusterPolicyBindingList{} },
-		DefaultQualifiedResource: authorizationapi.Resource("clusterpolicybindings"),
+		NewFunc:                  func() runtime.Object { return &authorizationapi.ClusterPolicy{} },
+		NewListFunc:              func() runtime.Object { return &authorizationapi.ClusterPolicyList{} },
+		DefaultQualifiedResource: authorizationapi.Resource("clusterpolicies"),
 
-		CreateStrategy: clusterpolicybinding.Strategy,
-		UpdateStrategy: clusterpolicybinding.Strategy,
-		DeleteStrategy: clusterpolicybinding.Strategy,
+		CreateStrategy: clusterpolicy.Strategy,
+		UpdateStrategy: clusterpolicy.Strategy,
+		DeleteStrategy: clusterpolicy.Strategy,
 	}
 
 	options := &generic.StoreOptions{RESTOptions: optsGetter}
