+++ pkg/cmd/server/admin/legacyetcd/clusterpolicy/registry.go
@@ -10,7 +10,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	"github.com/openshift/origin/pkg/authorization/registry/policy"
+	"github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/policy"
 )
 
 // Registry is an interface for things that know how to store ClusterPolicies.
