+++ pkg/cmd/server/admin/legacyetcd/clusterpolicybinding/etcd/etcd.go
@@ -7,7 +7,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	"github.com/openshift/origin/pkg/authorization/registry/policy"
+	"github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterpolicybinding"
 	"github.com/openshift/origin/pkg/util/restoptions"
 )
 
@@ -15,17 +15,17 @@ type REST struct {
 	*registry.Store
 }
 
-// NewREST returns a RESTStorage object that will work against Policy objects.
+// NewREST returns a RESTStorage object that will work against ClusterPolicyBinding.
 func NewREST(optsGetter restoptions.Getter) (*REST, error) {
 	store := &registry.Store{
 		Copier:                   kapi.Scheme,
-		NewFunc:                  func() runtime.Object { return &authorizationapi.Policy{} },
-		NewListFunc:              func() runtime.Object { return &authorizationapi.PolicyList{} },
-		DefaultQualifiedResource: authorizationapi.Resource("policies"),
+		NewFunc:                  func() runtime.Object { return &authorizationapi.ClusterPolicyBinding{} },
+		NewListFunc:              func() runtime.Object { return &authorizationapi.ClusterPolicyBindingList{} },
+		DefaultQualifiedResource: authorizationapi.Resource("clusterpolicybindings"),
 
-		CreateStrategy: policy.Strategy,
-		UpdateStrategy: policy.Strategy,
-		DeleteStrategy: policy.Strategy,
+		CreateStrategy: clusterpolicybinding.Strategy,
+		UpdateStrategy: clusterpolicybinding.Strategy,
+		DeleteStrategy: clusterpolicybinding.Strategy,
 	}
 
 	options := &generic.StoreOptions{RESTOptions: optsGetter}
