+++ pkg/cmd/server/admin/legacyetcd/clusterpolicybinding/registry.go
@@ -10,7 +10,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	"github.com/openshift/origin/pkg/authorization/registry/policybinding"
+	"github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/policybinding"
 )
 
 // Registry is an interface for things that know how to store ClusterPolicyBindings.
