+++ pkg/cmd/server/admin/legacyetcd/clusterrole/proxy/proxy.go
@@ -8,11 +8,11 @@ import (
 	"k8s.io/apiserver/pkg/registry/rest"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	clusterpolicyregistry "github.com/openshift/origin/pkg/authorization/registry/clusterpolicy"
 	"github.com/openshift/origin/pkg/authorization/registry/clusterrole"
 	roleregistry "github.com/openshift/origin/pkg/authorization/registry/role"
-	rolestorage "github.com/openshift/origin/pkg/authorization/registry/role/policybased"
 	"github.com/openshift/origin/pkg/authorization/rulevalidation"
+	clusterpolicyregistry "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterpolicy"
+	rolestorage "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/role/policybased"
 )
 
 type ClusterRoleStorage struct {
