+++ pkg/cmd/server/admin/legacyetcd/policy/etcd/etcd.go
@@ -7,7 +7,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	"github.com/openshift/origin/pkg/authorization/registry/clusterpolicy"
+	"github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/policy"
 	"github.com/openshift/origin/pkg/util/restoptions"
 )
 
@@ -15,17 +15,17 @@ type REST struct {
 	*registry.Store
 }
 
-// NewREST returns a RESTStorage object that will work against ClusterPolicy.
+// NewREST returns a RESTStorage object that will work against Policy objects.
 func NewREST(optsGetter restoptions.Getter) (*REST, error) {
 	store := &registry.Store{
 		Copier:                   kapi.Scheme,
-		NewFunc:                  func() runtime.Object { return &authorizationapi.ClusterPolicy{} },
-		NewListFunc:              func() runtime.Object { return &authorizationapi.ClusterPolicyList{} },
-		DefaultQualifiedResource: authorizationapi.Resource("clusterpolicies"),
+		NewFunc:                  func() runtime.Object { return &authorizationapi.Policy{} },
+		NewListFunc:              func() runtime.Object { return &authorizationapi.PolicyList{} },
+		DefaultQualifiedResource: authorizationapi.Resource("policies"),
 
-		CreateStrategy: clusterpolicy.Strategy,
-		UpdateStrategy: clusterpolicy.Strategy,
-		DeleteStrategy: clusterpolicy.Strategy,
+		CreateStrategy: policy.Strategy,
+		UpdateStrategy: policy.Strategy,
+		DeleteStrategy: policy.Strategy,
 	}
 
 	options := &generic.StoreOptions{RESTOptions: optsGetter}
