+++ pkg/cmd/server/admin/overwrite_bootstrappolicy.go
@@ -25,23 +25,23 @@ import (
 	templateapi "github.com/openshift/origin/pkg/template/apis/template"
 	"github.com/openshift/origin/pkg/util/restoptions"
 
-	clusterpolicyregistry "github.com/openshift/origin/pkg/authorization/registry/clusterpolicy"
-	clusterpolicyetcd "github.com/openshift/origin/pkg/authorization/registry/clusterpolicy/etcd"
-	clusterpolicybindingregistry "github.com/openshift/origin/pkg/authorization/registry/clusterpolicybinding"
-	clusterpolicybindingetcd "github.com/openshift/origin/pkg/authorization/registry/clusterpolicybinding/etcd"
 	"github.com/openshift/origin/pkg/authorization/registry/clusterrole"
-	clusterrolestorage "github.com/openshift/origin/pkg/authorization/registry/clusterrole/proxy"
 	"github.com/openshift/origin/pkg/authorization/registry/clusterrolebinding"
-	clusterrolebindingstorage "github.com/openshift/origin/pkg/authorization/registry/clusterrolebinding/proxy"
-	policyregistry "github.com/openshift/origin/pkg/authorization/registry/policy"
-	policyetcd "github.com/openshift/origin/pkg/authorization/registry/policy/etcd"
-	policybindingregistry "github.com/openshift/origin/pkg/authorization/registry/policybinding"
-	policybindingetcd "github.com/openshift/origin/pkg/authorization/registry/policybinding/etcd"
 	"github.com/openshift/origin/pkg/authorization/registry/role"
-	rolestorage "github.com/openshift/origin/pkg/authorization/registry/role/policybased"
 	"github.com/openshift/origin/pkg/authorization/registry/rolebinding"
-	rolebindingstorage "github.com/openshift/origin/pkg/authorization/registry/rolebinding/policybased"
 	"github.com/openshift/origin/pkg/authorization/rulevalidation"
+	clusterpolicyregistry "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterpolicy"
+	clusterpolicyetcd "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterpolicy/etcd"
+	clusterpolicybindingregistry "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterpolicybinding"
+	clusterpolicybindingetcd "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterpolicybinding/etcd"
+	clusterrolestorage "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterrole/proxy"
+	clusterrolebindingstorage "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/clusterrolebinding/proxy"
+	policyregistry "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/policy"
+	policyetcd "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/policy/etcd"
+	policybindingregistry "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/policybinding"
+	policybindingetcd "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/policybinding/etcd"
+	rolestorage "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/role/policybased"
+	rolebindingstorage "github.com/openshift/origin/pkg/cmd/server/admin/legacyetcd/rolebinding/policybased"
 	"github.com/openshift/origin/pkg/cmd/util/clientcmd"
 )
